Move document components to _components folder
