+++ src/features/ITGovernance/Requester/SystemIntake/Documents/index.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { useHistory } from 'react-router-dom';
 import { Button } from '@trussworks/react-uswds';
+import DocumentsTable from 'features/ITGovernance/_components/DocumentsTable';
 import Pager from 'features/TechnicalAssistance/Requester/RequestForm/Pager';
 import {
   SystemIntakeFormState,
@@ -12,8 +13,6 @@ import FeedbackBanner from 'components/FeedbackBanner';
 import PageHeading from 'components/PageHeading';
 import PageNumber from 'components/PageNumber';
 import useMessage from 'hooks/useMessage';
-
-import DocumentsTable from './DocumentsTable';
 
 import './index.scss';
